+++ src/design/plone/contenttypes/tests/test_relateditems_with_dates.py
@@ -30,9 +30,14 @@ class VocabulariesControlpanelTest(unittest.TestCase):
 
         self.api_session = RelativeSession(self.portal_url)
         self.api_session.headers.update({"Accept": "application/json"})
+        self.api_session_anon = RelativeSession(self.portal_url)
+        self.api_session_anon.headers.update({"Accept": "application/json"})
+
         self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
 
         setRoles(self.portal, TEST_USER_ID, ["Manager"])
+
+        self.intids = getUtility(IIntIds)
 
         published_news = api.content.create(
             container=self.portal, type="News Item", title="Published news"
@@ -52,21 +57,22 @@ class VocabulariesControlpanelTest(unittest.TestCase):
             end=datetime.now(),
         )
 
-        intids = getUtility(IIntIds)
         self.page = api.content.create(
             container=self.portal,
             type="Document",
             title="Page",
             relatedItems=[
-                RelationValue(intids.getId(published_news)),
-                RelationValue(intids.getId(private_news)),
-                RelationValue(intids.getId(event)),
+                RelationValue(self.intids.getId(published_news)),
+                RelationValue(self.intids.getId(private_news)),
+                RelationValue(self.intids.getId(event)),
             ],
         )
         commit()
 
     def tearDown(self):
         self.api_session.close()
+        self.api_session_anon.close()
+        setRoles(self.portal, TEST_USER_ID, ["Manager"])
 
     def test_api_do_not_return_effective_on_private_items(self):
         response = self.api_session.get(self.page.absolute_url())
@@ -101,3 +107,86 @@ class VocabulariesControlpanelTest(unittest.TestCase):
         # event
         self.assertIn("start", relatedItems[2])
         self.assertIn("end", relatedItems[2])
+
+    def test_api_do_not_return_related_items_with_effective_date_in_future_for_anon(
+        self,
+    ):
+
+        present = api.content.create(
+            container=self.portal, type="Document", title="present"
+        )
+        future = api.content.create(
+            container=self.portal, type="Document", title="future"
+        )
+        present.setEffectiveDate(DateTime())
+        future.setEffectiveDate(DateTime() + 1)
+        api.content.transition(obj=present, transition="publish")
+        api.content.transition(obj=future, transition="publish")
+        page = api.content.create(
+            container=self.portal,
+            type="Document",
+            title="Page",
+            relatedItems=[
+                RelationValue(self.intids.getId(present)),
+                RelationValue(self.intids.getId(future)),
+            ],
+        )
+        api.content.transition(obj=page, transition="publish")
+        commit()
+
+        res = self.api_session.get(page.absolute_url()).json()
+
+        relatedItems = res.get("relatedItems", [])
+        self.assertEqual(len(relatedItems), 2)
+
+        res_anon = self.api_session_anon.get(page.absolute_url()).json()
+        self.assertEqual(len(res_anon["relatedItems"]), 1)
+
+    def test_api_do_not_return_related_items_with_effective_date_in_future_for_users_that_cant_edit_context(
+        self,
+    ):
+        api.user.create(
+            email="foo@example.com",
+            username="foo",
+            password="secret",
+        )
+
+        api_session = RelativeSession(self.portal_url)
+        api_session.headers.update({"Accept": "application/json"})
+        api_session.auth = ("foo", "secret")
+
+        present = api.content.create(
+            container=self.portal, type="Document", title="present"
+        )
+        future = api.content.create(
+            container=self.portal, type="Document", title="future"
+        )
+        present.setEffectiveDate(DateTime())
+        future.setEffectiveDate(DateTime() + 1)
+        api.content.transition(obj=present, transition="publish")
+        api.content.transition(obj=future, transition="publish")
+        page = api.content.create(
+            container=self.portal,
+            type="Document",
+            title="Page",
+            relatedItems=[
+                RelationValue(self.intids.getId(present)),
+                RelationValue(self.intids.getId(future)),
+            ],
+        )
+        api.content.transition(obj=page, transition="publish")
+        commit()
+
+        setRoles(self.portal, "foo", ["Reader"])
+        commit()
+        res = api_session.get(page.absolute_url()).json()
+
+        relatedItems = res.get("relatedItems", [])
+        self.assertEqual(len(relatedItems), 1)
+
+        setRoles(self.portal, "foo", ["Editor"])
+        commit()
+        res = api_session.get(page.absolute_url()).json()
+
+        relatedItems = res.get("relatedItems", [])
+        self.assertEqual(len(relatedItems), 2)